backend/requirements.txt
aiofiles==24.1.0
aiomysql==0.2.0
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bidict==0.23.1
cffi==1.17.1
click==8.1.8
coverage==7.8.0
cryptography==42.0.5
defer==1.0.4
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.1.1
fastapi==0.115.12
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
joblib==1.4.2
jwt==1.3.1
Mako==1.3.9
MarkupSafe==3.0.2
nltk==3.8.1
numpy==1.26.4
packaging==24.2
pandas==2.2.1
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyMySQL==1.1.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-jose==3.3.0
python-multipart==0.0.9

pytz==2025.2
regex==2024.11.6
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.46.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
uvicorn==0.29.0